Backend/APIMobeen/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
audioread==3.0.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
decorator==5.1.1
dnspython==2.7.0
fastapi==0.115.0
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.25.1
HyperPyYAML==1.2.2
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
jsonify==0.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
MarkupSafe==3.0.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
packaging==24.1
platformdirs==4.3.6
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.1.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
starlette==0.38.6
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
transformers==4.45.2
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
zope.interface==7.0.3
